+++ src/app/cart/models/cart-item.model.ts
@@ -2,4 +2,5 @@ import { ProductModel } from "../../products/models/product.model";
 
 export interface CartItemModel extends ProductModel {
   readonly count: number;
+  readonly sum: number;
 }