+++ src/app/orders/orders.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+
+@NgModule({
+  declarations: [],
+  imports: [BrowserModule],
+  exports: [],
+  providers: [],
+})
+export class OrdersModule { }