+++ src/app/products/components/product/product.component.ts
@@ -1,15 +1,18 @@
-import { Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { ProductModel } from '../../models/product.model';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
-  styleUrls: ['./product.component.scss']
+  styleUrls: ['./product.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProductComponent {
   @Input() product!: ProductModel;
-
+  @Output() add = new EventEmitter<ProductModel>();
+  
   onAddToCart(): void {
+    this.add.emit(this.product);
     console.log("Item has been added!");
   }
 }