+++ src/app/products/products.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { ProductListComponent } from './components/product-list/product-list.component';
+import { ProductComponent } from './components/product/product.component';
+
+@NgModule({
+  declarations: [
+    ProductComponent,
+    ProductListComponent,
+  ],
+  imports: [
+    BrowserModule,
+  ],
+  exports:[
+    ProductComponent,
+    ProductListComponent,
+  ],
+  providers: [],
+})
+export class ProductsModule { }