+++ src/app/shared/shared.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { HighlightDirective } from './highlight.directive';
+
+@NgModule({
+  declarations: [
+    HighlightDirective
+  ],
+  imports: [],
+  exports: [HighlightDirective],
+  providers: [],
+})
+export class SharedModule { }